pyproject.toml
[tool.poetry]
name = "policymesh"
version = "0.1.0"
description = ""
authors = ["BloodCharry <vladvorobei1994@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.128.0,<0.129.0"
uvicorn = { version = ">=0.40.0,<0.41.0", extras = ["standard"] }
sqlalchemy = ">=2.0.45,<3.0.0"
asyncpg = ">=0.31.0,<0.32.0"
python-jose = { version = ">=3.5.0,<4.0.0", extras = ["cryptography"] }
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
python-multipart = ">=0.0.21,<0.0.22"
alembic = ">=1.18.1,<2.0.0"
email-validator = ">=2.3.0,<3.0.0"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
bcrypt = "^5.0.0"
pre-commit = "^4.5.1"
gunicorn = "^23.0.0"
structlog = "^25.5.0"
prometheus-fastapi-instrumentator = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
httpx = ">=0.28.1,<0.29.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
pre-commit = ">=4.5.1,<5.0.0"
black = ">=25.12.0,<26.0.0"
ruff = ">=0.14.11,<0.15.0"
mypy = ">=1.19.1,<2.0.0"
types-python-jose = "^3.5.0.20250531"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Black
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  | migrations
  | alembic
)/
'''

# Ruff
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501",
    "B008",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

# MyPy
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
plugins = ["pydantic.mypy"]
follow_imports = "silent"
mypy_path = ["."]
explicit_package_bases = true
